clean up maze layouts, define randMaze envs
