+++ d4rl/pointmaze/__init__.py
@@ -1,4 +1,6 @@
-from .maze_model import MazeEnv, OPEN, U_MAZE, MEDIUM_MAZE, LARGE_MAZE, HARD_EXP_MAZE, HARD_EXP_MAZE_V2
+from .maze_layouts import OPEN, U_MAZE, MEDIUM_MAZE, LARGE_MAZE, HARD_EXP_MAZE, HARD_EXP_MAZE_V2, rand_layout
+from .maze_model import MazeEnv
+
 from gym.envs.registration import register
 
 register(
@@ -104,6 +106,57 @@ register(
     }
 )
 
+
+register(
+    id='maze2d-randMaze0-ac-v0',
+    entry_point='d4rl.pointmaze:MazeEnv',
+    max_episode_steps=800,
+    kwargs={
+        'maze_spec': rand_layout(seed=0),
+        'agent_centric_view': True,
+        'reward_type': 'sparse',
+        'reset_target': False,
+        'ref_min_score': 4.83,
+        'ref_max_score': 191.99,
+        'dataset_url':'http://maze2d-hardexpv2-sparse.hdf5'
+    }
+)
+
+
+register(
+    id='maze2d-randMaze1-ac-v0',
+    entry_point='d4rl.pointmaze:MazeEnv',
+    max_episode_steps=800,
+    kwargs={
+        'maze_spec': rand_layout(seed=1),
+        'agent_centric_view': True,
+        'reward_type': 'sparse',
+        'reset_target': False,
+        'ref_min_score': 4.83,
+        'ref_max_score': 191.99,
+        'dataset_url':'http://maze2d-hardexpv2-sparse.hdf5'
+    }
+)
+
+
+register(
+    id='maze2d-randMaze42-ac-v0',
+    entry_point='d4rl.pointmaze:MazeEnv',
+    max_episode_steps=800,
+    kwargs={
+        'maze_spec': rand_layout(seed=42),
+        'agent_centric_view': True,
+        'reward_type': 'sparse',
+        'reset_target': False,
+        'ref_min_score': 4.83,
+        'ref_max_score': 191.99,
+        'dataset_url':'http://maze2d-hardexpv2-sparse.hdf5'
+    }
+)
+
+
+
+##################### OLD LAYOUTS #######################
 
 register(
     id='maze2d-open-dense-v0',
